Module amazon_sp_fulfillment_inbound::apis::fba_inbound_api

Enums

ConfirmPreorderError

struct for typed errors of method confirm_preorder

ConfirmTransportError

struct for typed errors of method confirm_transport

CreateInboundShipmentError

struct for typed errors of method create_inbound_shipment

CreateInboundShipmentPlanError

struct for typed errors of method create_inbound_shipment_plan

EstimateTransportError

struct for typed errors of method estimate_transport

GetBillOfLadingError

struct for typed errors of method get_bill_of_lading

GetInboundGuidanceError

struct for typed errors of method get_inbound_guidance

GetLabelsError

struct for typed errors of method get_labels

GetPreorderInfoError

struct for typed errors of method get_preorder_info

GetPrepInstructionsError

struct for typed errors of method get_prep_instructions

GetShipmentItemsByShipmentIdError

struct for typed errors of method get_shipment_items_by_shipment_id

GetShipmentItemsError

struct for typed errors of method get_shipment_items

GetShipmentsError

struct for typed errors of method get_shipments

GetTransportDetailsError

struct for typed errors of method get_transport_details

PutTransportDetailsError

struct for typed errors of method put_transport_details

UpdateInboundShipmentError

struct for typed errors of method update_inbound_shipment

VoidTransportError

struct for typed errors of method void_transport

Functions

confirm_preorder

Returns information needed to confirm a shipment for pre-order. Call this operation after calling the getPreorderInfo operation to get the NeedByDate value and other pre-order information about the shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

confirm_transport

Confirms that the seller accepts the Amazon-partnered shipping estimate, agrees to allow Amazon to charge their account for the shipping cost, and requests that the Amazon-partnered carrier ship the inbound shipment. Prior to calling the confirmTransport operation, you should call the getTransportDetails operation to get the Amazon-partnered shipping estimate. Important: After confirming the transportation request, if the seller decides that they do not want the Amazon-partnered carrier to ship the inbound shipment, you can call the voidTransport operation to cancel the transportation request. Note that for a Small Parcel shipment, the seller has 24 hours after confirming a transportation request to void the transportation request. For a Less Than Truckload/Full Truckload (LTL/FTL) shipment, the seller has one hour after confirming a transportation request to void it. After the grace period has expired the seller’s account will be charged for the shipping cost. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_inbound_shipment

Returns a new inbound shipment based on the specified shipmentId that was returned by the createInboundShipmentPlan operation. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_inbound_shipment_plan

Returns one or more inbound shipment plans, which provide the information you need to create one or more inbound shipments for a set of items that you specify. Multiple inbound shipment plans might be required so that items can be optimally placed in Amazon’s fulfillment network—for example, positioning inventory closer to the customer. Alternatively, two inbound shipment plans might be created with the same Amazon fulfillment center destination if the two shipment plans require different processing—for example, items that require labels must be shipped separately from stickerless, commingled inventory. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

estimate_transport

Initiates the process of estimating the shipping cost for an inbound shipment by an Amazon-partnered carrier. Prior to calling the estimateTransport operation, you must call the putTransportDetails operation to provide Amazon with the transportation information for the inbound shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_bill_of_lading

Returns a bill of lading for a Less Than Truckload/Full Truckload (LTL/FTL) shipment. The getBillOfLading operation returns PDF document data for printing a bill of lading for an Amazon-partnered Less Than Truckload/Full Truckload (LTL/FTL) inbound shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_inbound_guidance

Returns information that lets a seller know if Amazon recommends sending an item to a given marketplace. In some cases, Amazon provides guidance for why a given SellerSKU or ASIN is not recommended for shipment to Amazon’s fulfillment network. Sellers may still ship items that are not recommended, at their discretion. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_labels

Returns package/pallet labels for faster and more accurate shipment processing at the Amazon fulfillment center. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_preorder_info

Returns pre-order information, including dates, that a seller needs before confirming a shipment for pre-order. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_prep_instructions

Returns labeling requirements and item preparation instructions to help prepare items for shipment to Amazon’s fulfillment network. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_shipment_items

Returns a list of items in a specified inbound shipment, or a list of items that were updated within a specified time frame. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_shipment_items_by_shipment_id

Returns a list of items in a specified inbound shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_shipments

Returns a list of inbound shipments based on criteria that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_transport_details

Returns current transportation information about an inbound shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

put_transport_details

Sends transportation information to Amazon about an inbound shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

update_inbound_shipment

Updates or removes items from the inbound shipment identified by the specified shipment identifier. Adding new items is not supported. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

void_transport

Cancels a previously-confirmed request to ship an inbound shipment using an Amazon-partnered carrier. To be successful, you must call this operation before the VoidDeadline date that is returned by the getTransportDetails operation. Important: The VoidDeadline date is 24 hours after you confirm a Small Parcel shipment transportation request or one hour after you confirm a Less Than Truckload/Full Truckload (LTL/FTL) shipment transportation request. After the void deadline passes, your account will be charged for the shipping cost. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.