Function amazon_sp_fulfillment_inbound::apis::fba_inbound_api::get_bill_of_lading

pub async fn get_bill_of_lading(
    configuration: &Configuration,
    shipment_id: &str
) -> Result<GetBillOfLadingResponse, Error<GetBillOfLadingError>>

Returns a bill of lading for a Less Than Truckload/Full Truckload (LTL/FTL) shipment. The getBillOfLading operation returns PDF document data for printing a bill of lading for an Amazon-partnered Less Than Truckload/Full Truckload (LTL/FTL) inbound shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.