Function amazon_sp_fulfillment_inbound::apis::fba_inbound_api::create_inbound_shipment_plan

pub async fn create_inbound_shipment_plan(
    configuration: &Configuration,
    body: CreateInboundShipmentPlanRequest
) -> Result<CreateInboundShipmentPlanResponse, Error<CreateInboundShipmentPlanError>>

Returns one or more inbound shipment plans, which provide the information you need to create one or more inbound shipments for a set of items that you specify. Multiple inbound shipment plans might be required so that items can be optimally placed in Amazon’s fulfillment network—for example, positioning inventory closer to the customer. Alternatively, two inbound shipment plans might be created with the same Amazon fulfillment center destination if the two shipment plans require different processing—for example, items that require labels must be shipped separately from stickerless, commingled inventory. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.